Completes ex 7.2, adds view for single anectote
This commit adds a Route for a single anecdote, accessed at
'anecdotes/:id'. The main page of anecdotes now lists each anecdote
as a link to the single anecdote view.

The single anecdote view is a new Anecdote component defined in
App.js, which takes one anecdote object as a prop. The anecdote to
pass is identified using the useMatch hook. As a result, main.jsx is
modified so that the Router tags are in the main render around the
App component, and the App component returns a <div>.