Completes ex 7.3, navigates to '/' on create
This commit modifies the behavior on creating a new anecdote, so that
the application navigates back to the anecdote list at '/' upon
creating a new anecdote. It also adds a notification line below the
navigation menu for 5 seconds alerting the creation of the anecdote.
The notification is handled with react useReducer, and the code is
copied from the Part 6 Query Anecdotes project, available on github
at https://github.com/djanyreason/FSO_06/tree/main/query-anecdotes