Completes ex 7.6, eliminates invalid prop warning
This commit eliminates the console warning for passing an invalid
value for the reset prop in the Input HTML elements in the create new
anecdotes form. This is done by destructuring the variables created
with the useForm custom hook, separating off the reset prop into its
own variable and using spread to collect the remainder into an object
that can be passed as props to the Input field.

For purposes of full disclosure: this solution was taken directly
from a stack overflow post which I have linked below - based on the
sample code, it looks like the original question was posted by
someone doing this exact exercise

https://stackoverflow.com/questions/73930529/invalid-value-for-prop-reset-on-input-tag