Completes ex. 7.7, implements useCountry hook
This commit implements the useCountry custom hook so that it fetches
the country data from the API when the user presses the find button.
The useCountry hook makes use of useEffect to only make an API call
when the find button is clicked, which updates the 'name' parameter
passed to the hook. This update also makes minor changes to the
Country component to align with the data source's format.

As a note: the useCountry hook is written such that the 'not
found...' message does not display if the user clicks the find
button with an empty search bar. This is done so that it is not
displayed on the initial render - useCountry does not make API calls
with a blank name parameter and returns null. The exercise emphasizes
using useEffect's second parameter to control the updates, so the
natural solutions to this issue seem out of the scope of the exercise