Fixes 'not found...' on blank search issue
Immediately after the previous commit, I figured out a solution to
the issue I wrote about (at length) on how to distinguish the initial
render from subsequent searches with a blank search box within the
scope of the exercise.

This commit changes the if statement inside the useEffect block in
the useCountry custom hook to be executed if either the name
parameter is not blank or the country state variable is not null. As
the country state variable is updated after the first update to the
name parameter, this is able to distinguish between the initial
render and most later searches with blank strings.

The one case in which this still does not work on submitting blank
strings is if the user clicks 'find' after the first render before
with an empty input box - as this does not update 'name', it does not
trigger another execution of the useEffect block. I believe any
solution to this would be outside of the scope of the exercise, for
the reasons discussed in the previous commit. However, based on
recent history, if there is an elegant solution I'm sure it will come
to me shortly after I push this commit.