Completes ex. 7.8, implements useResource hook
This commit implements the useResource custom hook for the ultimate
hooks project. The useResource hook takes a base URL and returns an
object consisting of an array corresponding to the DB at that URL,
and an object containing a 'create' function which can be used to add
an entry to that database.