Adds project folders for Redux and React Query
FSO exercises 7.10-7.13 can be done in either Redux or React Query.
FSO notes state:

  If you want to maximize your learning, you should do both versions!

Accordingly, I plan to do both, and have seperate folders in this
repo for both versions.

In addition, this commit completes ex 7.9, which is to set up the
Prettier opinionated code formatter in my local installation of
VSCode. I don't believe completing this exercise results in any
changes to this repository.