Completes ex 7.11 w/ redux, refactor blogs info
This commit adds the information regarding blogs in the blog list to
the Redux store. It adds a reducer for the blogs store, which exports
two functions, one to initialize blogs from the database server, and
the other to add a new blog to the backend database. These functions
both handle communications with the server and manage the state of
the blogs store. The App, Newblog, and Bloglist components are all
updated to reflect this refactoring.