Completes ex 7.12 w/ redux, enables like & delete
This commit re-enables the like and delete functionality in the
bloglist application using redux to manage the bloglist. It adds two
new functions in the blogs reducer, one for liking a blog and one for
deleting a blog, both of which communicate with the backend to make
the update and, if successful, update the blog store. The commit also
refactors the App and Blog components to use these new functions.

This commit also adjusts the Bloglist component to employ reduxjs
toolkit's createSelector to avoid unnecessary sorts on re-renders.