Completes ex 7.13 w/ redux, moves login to store
This commit adds the state of a logged in user into the redux store,
with a new reducer userReducer. The userReducer file exports the
reducer along with two functions, setUser which takes a user object
and places it into the store, and logout which resets the user object
to null. Setting the user credentials in local storage is not handled
in the reducer, but instead by the react components. Components are
updated as necessary to implement or take advantage of this change.