Completes ex 7.12 w/ query, enables like & delete
This commit re-enables the like and delete functionality in the
bloglist application using query to manage the bloglist and backend
communication. It refactors the Blog component to use mutation of the
query client for the bloglist to update both the database and the
application state. Functions in the blogs service are updated to no
longer use async/await. the App and Bloglist component are updated to
enable and take advantage of the changes made in the Blog component.