Completes ex 7.13 w/ query, does login w/ Context
This commit uses useReducer and context to manage the data for a
logged in user. The context is defined in Contexts/UserContext and is
included as a wrapper component in main.jsx. The Login component is
modified to perform the login using this context, and the App
component is modified to perform logout using this context. App and
other components are futher modified to enable and take advantage of
this change.

Setting the user credentials in local storage is not handled
in the context, but instead by the react components.