Completes ex 7.14, adds routing and users view
This commit adds routing via React Router. It adds a new Menu
component which contains links to the blog list view (at '/') and to
a new user list view (at 'users'). The user list view presents a
table with the name of each user and the number of blogs they have
created.

To support the user list, /services/users.js is added to handle
communications with the users database. In addition, the former
userReducer is renamed to loginReducer, and a new userReducer is
added to the Redux store to manage the list of users.