Completes ex 7.15, adds inidividual user view
This commit adds a new User component which displays information
about an individual user. The view displays the user's name as a
header and an unordered list of the blogs they have added.

This view is Routed to 'users/:id', and linked to from the Userlist
component, with each user's name in the Userlist a link to their
User page.

This commit also updates userReducer to add a new reducer to update a
user, and two new functions addBlogToUser and removeBlogFromUser,
which are called when a new blog is created and when a blog is
deleted respectively. This ensures that the number of blogs displayed
in the Userlist component and list of blogs in the User component are
updated as changes are made to the bloglist app. To facilitate this,
a small change is made to the return object for the blogReducer
function createBlog.

Finally, an update is made to the Blog component to use the login and
users states in the Redux store to determine whether to display the
delete button, instead of receiving a boolean prop with this info.