Completes ex 7.16, adds individual blog view
This commit creates a view for an individual blog which is routed to
'/blogs/:id'. These pages are linked from the bloglist - the bloglist
now consists of links for blog names, and no longer has expandable
info, as that info has migrated to the individual blog pages.

The Blog component now takes no props, and selects the blog from the
store based on the id of the URL. When a blog is deleted from its
individual view, the bloglist navigates back to '/'.