Completes ex 7.18, adds comments to blog details
This commit adds a list of comments to the single blog detail view.
This is rendered via a new BlogComments component, which takes the
blog object as a prop. The former Blog component has been renamed to
BlogDetails, and is now also passed the blog object as a prop rather
than identifying it inside the component. A new Blog component is
added which identifies the blog obejct from the URL, and renders both
the BlogDetails and BlodComments components.

A few additional cleanup changes were made, removing imports which
are no longer used from various components.