Completes ex. 7.19, adds 'add comment' form
This commit enables adding comments via the frontend through a form
in the BlogComments component. Adding the comment to the backend is
handled through a new function added to services/blogs.js. A new
reducer is added to the blogsReducer to add the new comment to the
store, and a new function commentOnBlog is exported from
reducers/blogsReducer to make all the necessary calls to update the
store and the backend.